test(compat): relocate PreCutoverFailureFixture to shared Helpers.cs
